画面遷移追加 Class.js ShareRoom.js
